test: 깨지는 테스트 해결
