+++ tomcat/src/main/java/org/apache/coyote/http11/exception/UnauthorizeException.java
@@ -1,0 +1,7 @@
+package org.apache.coyote.http11.exception;
+
+public class UnauthorizeException extends RuntimeException {
+    public UnauthorizeException(String message) {
+        super(message);
+    }
+}
